pypbe-sim/requirements.txt
numpy
seaborn
matplotlib
pypbe
pyinstaller
pyqt5
pillow